test: update gc test
