+++ chain/index/ddls_test.go
@@ -662,47 +662,3 @@ func verifyNonRevertedEventEntriesCount(t *testing.T, s *SqliteIndexer, tipsetKe
 	require.NoError(t, err)
 	require.Equal(t, expectedCount, count)
 }
-
-func insertTipsetMessage(t *testing.T, s *SqliteIndexer, ts tipsetMessage) int64 {
-	res, err := s.stmts.insertTipsetMessageStmt.Exec(ts.tipsetKeyCid, ts.height, ts.reverted, ts.messageCid, ts.messageIndex)
-	require.NoError(t, err)
-
-	rowsAffected, err := res.RowsAffected()
-	require.NoError(t, err)
-	require.Equal(t, int64(1), rowsAffected)
-
-	messageID, err := res.LastInsertId()
-	require.NoError(t, err)
-	require.NotEqual(t, int64(0), messageID)
-
-	// read back the message to verify it was inserted correctly
-	verifyTipsetMessage(t, s, messageID, ts)
-
-	return messageID
-}
-
-func insertEvent(t *testing.T, s *SqliteIndexer, e event) int64 {
-	res, err := s.stmts.insertEventStmt.Exec(e.messageID, e.eventIndex, e.emitterAddr, e.reverted)
-	require.NoError(t, err)
-
-	rowsAffected, err := res.RowsAffected()
-	require.NoError(t, err)
-	require.Equal(t, int64(1), rowsAffected)
-
-	eventID, err := res.LastInsertId()
-	require.NoError(t, err)
-	require.NotEqual(t, int64(0), eventID)
-
-	verifyEvent(t, s, eventID, e)
-
-	return eventID
-}
-
-func insertEventEntry(t *testing.T, s *SqliteIndexer, ee eventEntry) {
-	res, err := s.stmts.insertEventEntryStmt.Exec(ee.eventID, ee.indexed, ee.flags, ee.key, ee.codec, ee.value)
-	require.NoError(t, err)
-
-	rowsAffected, err := res.RowsAffected()
-	require.NoError(t, err)
-	require.Equal(t, int64(1), rowsAffected)
-}
